Minimize header includes by moving impl to .cpp files (#331) and Fix usage of rclcpp::ok with a non-default context (#352)  (#390)
* Minimize header includes by moving impl to .cpp files (#331)

* Minimize header includes by moving impl to .cpp files
* Make sure to build a shared library

Signed-off-by: Christian Henkel <christian.henkel2@de.bosch.com>

* Fix usage of rclcpp::ok with a non-default context (#352)

* Fix usage of rclcpp::ok with a non-default context

The current implementation calls `rclcpp::ok` without any arguments,
which amounts to verifying that the global default context is valid. In
the case where a node is added to a custom context, and the global
context is not used, `rclcpp::ok` without any arguments will always
return `false` since the global context has never been initialized. To
fix it, pass to rclcpp the context that's available via the node's base
interface.

* Add a test for custom context

Signed-off-by: Christian Henkel <christian.henkel2@de.bosch.com>

---------

Signed-off-by: Christian Henkel <christian.henkel2@de.bosch.com>
Co-authored-by: Ramon Wijnands <ramon.wijnands007@gmail.com>
Co-authored-by: Hervé Audren <101862279+haudren-woven@users.noreply.github.com>